+++ src/snowflake/cli/plugins/connection/commands.py
@@ -312,7 +312,8 @@ def test(
 @app.command(requires_connection=False)
 def set_default(
     name: str = typer.Argument(
-        help="Name of the connection, as defined in your `config.toml`"
+        help="Name of the connection, as defined in your `config.toml`",
+        show_default=False,
     ),
     **options,
 ):