+++ src/snowflake/cli/templates/default_streamlit/streamlit_app.py
@@ -1,4 +1,0 @@
-import streamlit as st
-from common.hello import say_hello
-
-st.title(f"Example streamlit app. {say_hello()}")