+++ tests_integration/testing_utils/snowpark_utils.py
@@ -241,7 +241,7 @@ class SnowparkTestSteps:
 
     def package_should_build_proper_artifact(self, package_name: str, file_name: str):
         result = self._setup.runner.invoke_with_connection_json(
-            ["snowpark", "package", "create", package_name, "--pypi-download"]
+            ["snowpark", "package", "create", package_name]
         )
 
         assert result.exit_code == 0